refactor: view package 분리
