+++ src/main/java/study/step3/view/ResultView.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.step3.view;
 
 import study.step3.domain.CarRacing;
 import study.step3.domain.RaceRecord;
@@ -7,7 +7,7 @@ import java.util.*;
 
 import static java.util.stream.Collectors.joining;
 
-class ResultView {
+public class ResultView {
     private final RaceRecord raceRecord;
     private final Set<String> winners;
