+++ src/test/java/study/step3/view/InputViewTest.java
@@ -1,10 +1,11 @@
-package study.step3;
+package study.step3.view;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import study.step3.domain.Circuit;
 import study.step3.domain.RealCar;
+import study.step3.view.InputView;
 
 import java.util.ArrayDeque;
 import java.util.Queue;
